fix: 将保存到数据库中的价格改为原始价格*100，防止浮点数比较大小精度丢失

+++ api-model/src/main/java/xyz/linyh/model/pay/dto/CreditProductDto.java
@@ -28,6 +28,6 @@ public class CreditProductDto {
     /**
      * 折扣价格
      */
-    private Double discountPrice;
+    private Integer discountPrice;
 
 }
